# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SlugEntry {
  slug: string;
  page_type: string;
  region: string;
}

export interface PageLink {
  slug: string;
  title: string;
  h1_heading: string;
}

export interface PageDetail {
  id: number;
  page_type: string;
  slug: string;
  title: string;
  h1_heading: string;
  niche_key: string;
  intent_type: string;
  /** @nullable */
  target_city?: string | null;
  target_state: string;
  region: string;
  related_city_pages: PageLink[];
  related_state_pages: PageLink[];
}

export interface StateInfo {
  state_name: string;
  state_slug: string;
  region: string;
}

export interface NicheInfo {
  niche_key: string;
  display_name: string;
}

export interface ErrorResponse {
  error: string;
}