# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages/{slug}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a programmatic page by slug
      description: Returns full page data including related city and state links
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Page detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDetail"
        "404":
          description: Page not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    PageLink:
      type: object
      required: [slug, title, h1_heading]
      properties:
        slug:
          type: string
        title:
          type: string
        h1_heading:
          type: string
    PageDetail:
      type: object
      required: [id, page_type, slug, title, h1_heading, niche_key, intent_type, target_state, region, related_city_pages, related_state_pages]
      properties:
        id:
          type: integer
        page_type:
          type: string
        slug:
          type: string
        title:
          type: string
        h1_heading:
          type: string
        niche_key:
          type: string
        intent_type:
          type: string
        target_city:
          type: ["string", "null"]
        target_state:
          type: string
        region:
          type: string
        related_city_pages:
          type: array
          items:
            $ref: "#/components/schemas/PageLink"
        related_state_pages:
          type: array
          items:
            $ref: "#/components/schemas/PageLink"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string