# lib/api-zod/src/generated/types/pageDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PageLink } from './pageLink';

export interface PageDetail {
  id: number;
  page_type: string;
  slug: string;
  title: string;
  h1_heading: string;
  niche_key: string;
  intent_type: string;
  /** @nullable */
  target_city?: string | null;
  target_state: string;
  region: string;
  related_city_pages: PageLink[];
  related_state_pages: PageLink[];
}