# lib/api-zod/src/generated/types/pageLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PageLink {
  slug: string;
  title: string;
  h1_heading: string;
}